+++ source-code/clj/server_extensions/clients/client_lister/lifecycles.clj
@@ -3,7 +3,8 @@
 ;; ----------------------------------------------------------------------------
 
 (ns server-extensions.clients.client-lister.lifecycles
-    (:require [x.server-core.api :as a]))
+    (:require [server-plugins.item-lister.api]
+              [x.server-core.api :as a]))
 
 
 
@@ -12,7 +13,6 @@
 
 (a/reg-lifecycles!
   ::lifecycles
-  {:on-server-boot [:router/add-route! :clients.client-lister/route
-                                       {:route-template "/@app-home/clients"
-                                        :client-event   [:clients.client-lister/load-lister!]
-                                        :restricted?    true}]})
+  {:on-server-boot [:item-lister/init-lister! :clients :client
+                                              {:base-route "/@app-home/clients"
+                                               :on-load    [:clients.client-lister/load-lister!]}]})