+++ source-code/clj/server_extensions/storage/media_browser/lifecycles.clj
@@ -3,7 +3,8 @@
 ;; ----------------------------------------------------------------------------
 
 (ns server-extensions.storage.media-browser.lifecycles
-    (:require [server-extensions.storage.engine :as engine]
+    (:require [server-plugins.item-browser.api]
+              [server-extensions.storage.engine :as engine]
               [x.server-core.api                :as a]))
 
 
@@ -13,11 +14,6 @@
 
 (a/reg-lifecycles!
   ::lifecycles
-  {:on-server-boot {:dispatch-n [[:router/add-route! :storage.media-browser/route
-                                                     {:route-template "/@app-home/storage"
-                                                      :client-event [:storage.media-browser/load-browser!]
-                                                      :restricted? true}]
-                                 [:router/add-route! :storage.media-browser/extended-route
-                                                     {:route-template "/@app-home/storage/:media-id"
-                                                      :client-event [:storage.media-browser/load-browser!]
-                                                      :restricted? true}]]}})
+  {:on-server-boot [:item-browser/init-browser! :storage :media
+                                                {:base-route "/@app-home/storage"
+                                                 :on-load    [:storage.media-browser/load-browser!]}]})