+++ source-code/cljs/x/app_ui/header/views.cljs
@@ -22,10 +22,10 @@
 (defn header-go-up-icon-button
   ; WARNING! NON-PUBLIC! DO NOT USE!
   []
-  (if-let [route-parent @(a/subscribe [:ui/get-route-parent])]
-          (case route-parent "/@app-home" [header-go-home-icon-button]
+  (if-let [parent-route @(a/subscribe [:ui/get-parent-route])]
+          (case parent-route "/@app-home" [header-go-home-icon-button]
                                           [elements/icon-button ::go-up-icon-button
-                                                                {:preset :back :on-click [:router/go-to! route-parent]}])))
+                                                                {:preset :back :on-click [:router/go-to! parent-route]}])))
 
 (defn header-at-home-icon-button
   ; WARNING! NON-PUBLIC! DO NOT USE!