+++ documentation/CHANGES.md
@@ -1,4 +1,14 @@
 
+# x4.6.5
+
+- A [:ui/set-route-parent! ...] esemény ezentúl [:ui/set-parent-route! ...] néven elérhető
+  (a pluginok szerver-oldali inicializálásakor átadott {:base-route ...} tulajdonság alapján automatikusan beállítódik)
+
+- Az item-lister plugin {:new-item-route ...} tulajdonsága helyett ezentúl {:new-item-event ...} tulajdonság használható
+
+
+
+
 # x4.6.4
 
 - Az útvonalak {:route-parent "..."} beállítása ezentúl nem elérhető, helyette a kliens-oldali